# Use jest instead of karma for running tests