--- a/bin/web
+++ b/bin/web
@@ -19,7 +19,7 @@ USAGE: web <command>
   * setup - get the environment setup for development
       Note: any command line options are passed on to yarn
   * test - run the tests
-      Note: any command line options are passed on to karma
+      Note: any command line options are passed on to jest
 USAGE
 }; function --help { -h ;}
 
@@ -81,7 +81,7 @@ function setup {
 
 function test {
   cd $ROOT/web/app
-  yarn karma start --single-run $*
+  yarn jest "$*"
 }
 
 function main {