--- a/web/app/__mocks__/fileMock.js
+++ b/web/app/__mocks__/fileMock.js
@@ -1,0 +1,1 @@
+module.exports = 'test-file-stub';