--- a/web/app/__mocks__/styleMock.js
+++ b/web/app/__mocks__/styleMock.js
@@ -1,0 +1,1 @@
+module.exports = {};