--- a/web/app/js/components/Version.test.jsx
+++ b/web/app/js/components/Version.test.jsx
@@ -1,9 +1,9 @@
 import Adapter from 'enzyme-adapter-react-16';
-import ApiHelpers from '../js/components/util/ApiHelpers.jsx';
+import ApiHelpers from './util/ApiHelpers.jsx';
 import { BrowserRouter } from 'react-router-dom';
 import { expect } from 'chai';
 import React from 'react';
-import Sidebar from '../js/components/Sidebar.jsx';
+import Sidebar from './Sidebar.jsx';
 import sinon from 'sinon';
 import sinonStubPromise from 'sinon-stub-promise';
 import Enzyme, { mount } from 'enzyme';