--- a/web/app/package.json
+++ b/web/app/package.json
@@ -9,6 +9,7 @@
     "d3": "^4.11.0",
     "enzyme": "^3.6.0",
     "enzyme-adapter-react-16": "^1.5.0",
+    "jest": "^23.6.0",
     "jsdom": "^11.5.1",
     "karma": "^1.7.1",
     "karma-jsdom-launcher": "^6.1.2",