--- a/web/app/package.json
+++ b/web/app/package.json
@@ -38,6 +38,7 @@
   "devDependencies": {
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.3",
+    "babel-jest": "^23.6.0",
     "babel-loader": "^7.1.4",
     "babel-plugin-import": "^1.7.0",
     "babel-preset-env": "^1.7.0",
@@ -53,6 +54,8 @@
     "file-loader": "^1.1.5",
     "history": "^4.7.2",
     "jest": "^23.6.0",
+    "jest-dot-reporter": "^1.0.7",
+    "jest-enzyme": "^7.0.0",
     "less": "^3.8.1",
     "less-loader": "^4.1.0",
     "less-vars-to-js": "^1.3.0",