--- a/web/app/package.json
+++ b/web/app/package.json
@@ -10,7 +10,6 @@
     "enzyme": "^3.6.0",
     "enzyme-adapter-react-16": "^1.5.0",
     "jsdom": "^11.5.1",
-    "karma": "^1.7.1",
     "karma-jsdom-launcher": "^6.1.2",
     "karma-mocha": "^1.3.0",
     "karma-nyan-reporter": "^0.2.5",