--- a/web/app/package.json
+++ b/web/app/package.json
@@ -11,7 +11,6 @@
     "enzyme-adapter-react-16": "^1.5.0",
     "jsdom": "^11.5.1",
     "karma-jsdom-launcher": "^6.1.2",
-    "karma-mocha": "^1.3.0",
     "karma-nyan-reporter": "^0.2.5",
     "karma-sinon-chai": "^1.3.3",
     "karma-webpack": "^2.0.6",