… web/app/karma.conf.js
… web/app/yarn.lock