Jest is faster, has more flexibility to run a subset of the tests, and will allow us to remove a bunch of our assertion libraries.

Many thanks to @grampelberg for prior work on this (#1000)

This PR:
- changes the test runner from karma to jest
- moves individual tests from `/test/ to `/js/components` where jest expects them

In a follow up PR, I will:
- remove the assertion libraries in favor of jest assertions (mocha, chai) - #1769
- further speed up tests by using shallow instead of mount - #1771
- lint the test files

Before:
```
$ NODE_ENV=test yarn karma start
yarn run v1.10.1
$ /Users/mars/workspace/go/src/github.com/linkerd/linkerd2/web/app/node_modules/.bin/karma start
(node:53371) DeprecationWarning: Tapable.plugin is deprecated. Use new API on `.hooks` instead
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
webpack: wait until bundle finished:
15 10 2018 12:49:05.247:WARN [karma]: No captured browser, open http://localhost:9876/
Hash: 0b00965308fc3d11c7ba
Version: webpack 4.19.1
Time: 16491ms
Built at: 10/15/2018 12:49:03 PM
```

After:
```
$ yarn jest "$*"
yarn run v1.10.1
$ /Users/mars/workspace/go/src/github.com/linkerd/linkerd2/web/app/node_modules/.bin/jest ''
 PASS  js/components/util/MetricUtils.test.js
 PASS  js/components/util/ApiHelpers.test.jsx
 PASS  js/components/Version.test.jsx
 PASS  js/components/Sidebar.test.jsx
 PASS  js/components/ServiceMesh.test.jsx (5.493s)
 PASS  js/components/MetricsTable.test.js
 PASS  js/components/ResourceList.test.jsx
 PASS  js/components/NetworkGraph.test.jsx
 PASS  js/components/BreadcrumbHeader.test.jsx
 PASS  js/components/GrafanaLink.test.jsx
 PASS  js/components/util/Utils.test.js

Test Suites: 11 passed, 11 total
Tests:       64 passed, 64 total
Snapshots:   0 total
Time:        7.805s, estimated 8s
Ran all test suites.
✨  Done in 9.14s.
```