@rmars 2018-10-15
Yeah! I think snapshots might be good to do. I made a ticket to investigate their use!